Switched min_heap to size_t to prevent integer overflows.
Also, as deraadt suggested, switched realloc to recallocarray to at least
prevent uninitialized memory to be used as pointers in case of other
programming errors. A proper solution (not using an array) needs more work.

This change occured in sync with upstream libevent 2.2.

with input by and ok bluhm, jca, tedu

+++ lib/libevent/event.h
@@ -1,4 +1,4 @@
-/*	$OpenBSD: event.h,v 1.30 2015/01/05 23:14:36 bluhm Exp $	*/
+/*	$OpenBSD: event.h,v 1.31 2019/04/29 17:11:51 tobias Exp $	*/
 
 /*
  * Copyright (c) 2000-2007 Niels Provos <provos@citi.umich.edu>
@@ -196,7 +196,7 @@ struct event {
 	TAILQ_ENTRY (event) ev_next;
 	TAILQ_ENTRY (event) ev_active_next;
 	TAILQ_ENTRY (event) ev_signal_next;
-	unsigned int min_heap_idx;	/* for managing timeouts */
+	size_t min_heap_idx;	/* for managing timeouts */
 
 	struct event_base *ev_base;
